Create Multigrid LS type, allow nonsquare petsc matrices
Created a new type for petsc multigrid linear solvers, and gave it
its own init function.  MatrixTypes_PETSc was modified to allow for
nonsquare sparse matrices.

I tried running it, but for some reason I'm getting a SEGV error when
KSPSolve is called.  I need to figure out why that is happening.