Use a matrix shell to define interp. operator
Per Jed Brown's suggestion, I rewrote the interpolation/restriction setting so
that the interpolation/restriction matrices are defined implicitly by their
operations on vectors.  I still need to add the PtAP operation for this to work.

I also made a clear() subroutine for LS_Multigrid to deal with clearing all the
variables specific to the multigrid type.  The unit test was updated to include
a test for the SetupInterpMats subroutine.