Make multigrid work for a simple 1D diffusion system
I fixed the dnnz and onnz inputs, which helped with the allocation of all the
problem and interpolation matrices.

For future reference:
dnnz(i) = # of nonzeros in row i local to the processor
onnz(i) = # of nonzeros in the row stored in other processors

I also fixed the setting of the smoothers on each level.  KSPRICHARDSON should
be used whenever we want a non-Krylov iterative solver that is stored in PETSc
as a preconditioner.