Refactor multigrid code
I realized that it would be too difficult to automate the generation of
interpolation matrices in Futility.  These matrices are too problem-dependent.
I've split up the generation of the interpolation matrices into three parts --
a preallocation part where the user must tell PETSc the sizes and nonzero
structure of each matrix, a part where the user must manually define the
matrices, and then a final part where the interpolation matrices are set in the
PCMG solver.