Add 2G test, fix LS_MG to work for 2G problem
A second 2-group 1-D problem was added to testLinearSolver_Multigrid.f90.
Supporting subroutines were added to generate the 2G problem and its
interpolation and restriction matrices.

The block Jacobi smoother was corrected in LinearSolverTypes_Multigrid.f90.
It does block Jacobi and each block is solved via Gauss-Seidel on the block.