Manually set PETSc convergence criteria for now.
Some test runs were hitting a maximum of 10000 iterations.  I limited them
to 10 and got effectively the same answer.  This indicates that I'm not
defining the convergence criteria in PETSc properly or that I don't understand
how it works.  In any case, I am limiting the max number of iterations for now.