Remove smoother subroutine, define 1D2G 2-proc unit test
I decided to remove the smoother subroutine.  In the PETSc context, it doesn't
really make sense since I need a specific subroutine that fits PETSc's shell
subroutine definitions.

In the overall scheme, there will be changes to the way the smoothers are
organized.  Namely, I will need to define a module-wide list of smoothers that
will be managed through public manager subroutines.