Significant refactor of code to wrap all smoother types with a manager
To use PCSHELL in petsc, we have to define several subroutines -- the most
important of which is PCApply, which takes in a vector x and returns a
preconditioned vector y.  This preconditioning (smoothing) step is where the
CBJ smoother will act on the preconditioner vector y.  However, it needs
access to the coloring schemes (and possibly the LU factorizations of the
blocks) which are/will be saved in the SmootherTypes.  Thus, a module-scope
list is needed, and this list is managed by various manager subroutines that
can be called by external modules.